Refactored Workflows import/export step.
